Split this into internal and public header
